Move storage events to an event subscriber class
