+++ src/EventListener/StorageListener.php
@@ -1,0 +1,227 @@
+<?php
+
+namespace Bolt\Extension\Animal\Translate\EventListener;
+
+use Bolt\Config;
+use Bolt\Events\HydrationEvent;
+use Bolt\Events\StorageEvent;
+use Bolt\Events\StorageEvents;
+use Bolt\Storage\Entity\Content;
+use Bolt\Storage\Field\Collection\RepeatingFieldCollection;
+use Bolt\Storage\Query\Query;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\HttpFoundation\RequestStack;
+
+/**
+ * Storage event subscriber.
+ *
+ * @author Svante Richter <svante.richter@gmail.com>
+ * @author Gawain Lynch <gawain.lynch@gmail.com>
+ */
+class StorageListener implements EventSubscriberInterface
+{
+    /** @var string */
+    protected $localeSlug;
+
+    /** @var Config */
+    private $boltConfig;
+    /** @var Query $query */
+    private $query;
+    /** @var RequestStack */
+    private $requestStack;
+    /** @var array */
+    private $config;
+
+    /**
+     * Constructor.
+     *
+     * @param Config       $boltConfig
+     * @param array        $config
+     * @param Query        $query
+     * @param RequestStack $requestStack
+     */
+    public function __construct(Config $boltConfig, array $config, Query $query, RequestStack $requestStack)
+    {
+        $this->boltConfig = $boltConfig;
+        $this->config = $config;
+        $this->query = $query;
+        $this->requestStack = $requestStack;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public static function getSubscribedEvents()
+    {
+        return [
+            StorageEvents::PRE_HYDRATE => [
+                ['preHydrate', 0],
+            ],
+            StorageEvents::POST_HYDRATE => [
+                ['postHydrate', 0],
+            ],
+            StorageEvents::PRE_SAVE => [
+                ['preSave', 0],
+            ],
+            StorageEvents::POST_SAVE => [
+                ['postSave', 0],
+            ],
+        ];
+    }
+
+    /**
+     * StorageEvents::PRE_HYDRATE event callback.
+     *
+     * @param HydrationEvent $event
+     */
+    public function preHydrate(HydrationEvent $event)
+    {
+        $request = $this->requestStack->getCurrentRequest();
+        if ($request === null) {
+            return;
+        }
+
+        /** @var Content $entity */
+        $entity = $event->getArgument('entity');
+        $subject = $event->getSubject();
+
+        if (!$entity instanceof Content || $request->query->getBoolean('no_locale_hydrate')) {
+            return;
+        }
+
+        $contentTypeName = $entity->getContenttype();
+        $contentType = $this->boltConfig->get('contenttypes/' . $contentTypeName);
+
+        if (isset($subject[$this->localeSlug . '_data'])) {
+            $localeData = json_decode($subject[$this->localeSlug . '_data'], true);
+            foreach ($localeData as $key => $value) {
+                if ($contentType['fields'][$key]['type'] !== 'repeater') {
+                    $subject[$key] = is_array($value) ? json_encode($value) : $value;
+                }
+            }
+        }
+    }
+
+    /**
+     * StorageEvents::POST_HYDRATE event callback.
+     *
+     * @param HydrationEvent $event
+     */
+    public function postHydrate(HydrationEvent $event)
+    {
+        $request = $this->requestStack->getCurrentRequest();
+        if ($request === null) {
+            return;
+        }
+
+        /** @var Content $subject */
+        $subject = $event->getSubject();
+        if (!$subject instanceof Content || $request->query->getBoolean('no_locale_hydrate')) {
+            return;
+        }
+
+        $contentTypeName = $subject->getContenttype();
+        $contentType = $this->boltConfig->get('contenttypes/' . $contentTypeName);
+
+        if (!isset($subject[$this->localeSlug . '_data'])) {
+            return;
+        }
+        $localeData = json_decode($subject[$this->localeSlug . '_data'], true);
+        foreach ($localeData as $key => $value) {
+            if ($contentType['fields'][$key]['type'] !== 'repeater') {
+                continue;
+            }
+            /** @var RepeatingFieldCollection[] $subject */
+            $subject[$key]->clear();
+            foreach ($value as $subValue) {
+                $subject[$key]->addFromArray($subValue);
+            }
+        }
+    }
+
+    /**
+     * StorageEvents::PRE_SAVE event callback.
+     *
+     * @param StorageEvent $event
+     */
+    public function preSave(StorageEvent $event)
+    {
+        $contentType = $this->boltConfig->get('contenttypes/' . $event->getContentType());
+        $translatableFields = $this->getTranslatableFields($contentType['fields']);
+        /** @var Content $record */
+        $record = $event->getContent();
+        $values = $record->serialize();
+        $localeValues = [];
+
+        if (empty($translatableFields)) {
+            return;
+        }
+
+        $record->set($this->localeSlug . '_slug', $values['slug']);
+        if ($values['locale'] == array_keys($this->config['locales'])[0]) {
+            $record->set($this->localeSlug . '_data', '[]');
+
+            return;
+        }
+
+        if ($values['id']) {
+            /** @var Content $defaultContent */
+            $defaultContent = $this->query->getContent(
+                $event->getContentType(),
+                ['id' => $values['id'], 'returnsingle' => true]
+            );
+        }
+        foreach ($translatableFields as $field) {
+            $localeValues[$field] = $values[$field];
+            if ($values['id']) {
+                $record->set($field, $defaultContent->get($field));
+            } else {
+                $record->set($field, '');
+            }
+        }
+        $localeJson = json_encode($localeValues);
+        $record->set($this->localeSlug . '_data', $localeJson);
+    }
+
+    /**
+     * StorageEvents::POST_SAVE event callback.
+     *
+     * @param StorageEvent $event
+     */
+    public function postSave(StorageEvent $event)
+    {
+        $subject = $event->getSubject();
+        if (!$subject instanceof Content) {
+            return;
+        }
+        if (isset($subject[$this->localeSlug . '_data'])) {
+            return;
+        }
+
+        $localeData = json_decode($subject[$this->localeSlug . '_data']);
+        foreach ($localeData as $key => $value) {
+            $subject->set($key, $value);
+        }
+    }
+
+    /**
+     * Helper to check for translatable fields in a contenttype
+     *
+     * @param array $fields
+     *
+     * @return array
+     */
+    private function getTranslatableFields($fields)
+    {
+        $translatable = [];
+        foreach ($fields as $name => $field) {
+            if (isset($field['is_translateable'])  && $field['is_translateable'] === true && $field['type'] === 'templateselect') {
+                $translatable[] = 'templatefields';
+            } elseif (isset($field['is_translateable']) && $field['is_translateable'] === true) {
+                $translatable[] = $name;
+            }
+        }
+
+        return $translatable;
+    }
+}
